+++ src/commands/symlink.js
@@ -142,7 +142,6 @@ const symlink = (target, path, ...args) => {
     const showPath = new SettingManager().fullOrBase(path);
 
     if (err.code === "EPERM" || err.code === "EACCES") {
-      // If there are no permissions to make the symbolic link
       // Note that on Windows (and maybe Linux/macOS), you need
       // to run it with elevated privileges to make the command work.
       Verbose.permError();